Replace ascii spaces with Japanese full-width spaces in subtitles.
This is a tweak pretty specifically for me: I really only use
Japanese subtitles, and this makes adding furigana less troublesome
after exporting to Anki.

A more robust implementation would check the language of the subtitles
first, but I don't care enough right now to take the time to figure
out how best to do that.